tableWithOdp: moved copy & paste utility functions to copyPasteUtils
